actually check the type of the underlying control rather than what we think it should be
For some reason, there are some cases where ImplementedAs() returns a different value than what it should for the given underlying control. I think this is because its concept of the difference between TreeView/ListView changes in the meantime. Will need to look into that; it's an underlying bug that needs to be fixed.